NEW - Se unificaron todas las clases en un mismo archivo gestor_datos.py con la intención de hacer más sencilla la importación
